- Changed uint32_t to size_t for calls to  GetBufferPointerUnsafe as this inconsistency caused  some compilers to fail.

+++ trunk/src/common/net/npcmessages.cpp
@@ -225,7 +225,7 @@ csString psNPCCommandsMessage::ToString(AccessPointers * access_ptrs)
                 msgtext.Append("CMD_DRDATA: ");
 
                 // Extract the data
-                uint32_t len = 0;
+                size_t len = 0;
 				void *data = msg->GetBufferPointerUnsafe(len);
 
                 // Make sure we haven't run past the end of the buffer
